+++ RELEASE-NOTES.md
@@ -25,6 +25,7 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 - Inlet: Added support for string-keyed associative arrays (dictionaries)
 - Inlet: Added support for defining and retrieving functions from the input file
 - Inlet: Added support for YAML and JSON input files
+- Inlet: Added support for mixed-key (integer and string) associative arrays
 
 ### Changed
 - The Sidre Datastore no longer rewires Conduit's error handlers to SLIC by default.